+++ lexxpluss_apps/src/message.hpp
@@ -46,8 +46,8 @@ struct msg_ros2actuator {
 
 struct msg_actuator2ros {
     int32_t encoder_count[3];
-    uint16_t current[3];
-    uint16_t connect;
+    int32_t current[3];
+    int32_t connect;
 } __attribute__((aligned(4)));
 
 struct msg_uss2ros {
